+++ CHANGELOG.md
@@ -4,6 +4,11 @@ PowerModels.jl Change Log
 ### Staged
 - nothing
 
+### v0.15.2
+- Add support for on/off storage in SOCWRPowerModel
+- Fixed multinetwork solution building bug
+- Fixed solution building bug in variable_demand_factor
+
 ### v0.15.1
 - Add support for JuMP v0.21